fix: contract types updated

+++ src/types/factories/ICreditFacadeV3.sol/ICreditFacadeV3__factory.ts
@@ -189,9 +189,9 @@ const _abi = [
       },
       {
         indexed: false,
-        internalType: "uint16",
+        internalType: "uint256",
         name: "referralCode",
-        type: "uint16",
+        type: "uint256",
       },
     ],
     name: "OpenCreditAccount",
@@ -600,9 +600,9 @@ const _abi = [
         type: "tuple[]",
       },
       {
-        internalType: "uint16",
+        internalType: "uint256",
         name: "referralCode",
-        type: "uint16",
+        type: "uint256",
       },
     ],
     name: "openCreditAccount",
